Implemented Root() and SetRoot() for a Container.
SetRoot() sets the root types.Type for the Container and, when a root
type has been set, Root() returns the root node for the Container. These
methods are intentionally not a getter/setter pair because the presense
of a root node is more useful then being able to retreive the types.Type
of that led to their being a root.