Simplified and implemented edgeImpl.
This struct now implments the graph.Edge interface but does not
implement the depend.edge interface. If depend.edge becomes necessary
then edgeImpl can be extended to support it. If it does not become
necessary then depend.edge should be removed (and the other interfaces
in package depend adjusted accordingly).