Changed and implemented require() and provide() for nodes.
The node interface that has require() and provide() now returns a slice
of types.Types instead of a slice of edge interfaces. The edge interface
is going to be removed.

missingNode does not requires() or provides() anything. This is different
from the edges that missingNode node may have in a container.

rootNode does not provides() anything but does requires() the root type set
on the Container by SetRoot(). This correspond to the edge comming into
the rootNode.

funcNode requires() the types of the paramters to its types.Func. It
provides() the non-error return types of its types.Func. These
correspond to the incoming and outgoing edges of the funcNode
respectively.