Initial implementation of Container.AddFunc.
The initial part of the implementation screens the function to be added
for it being a method or for having an error return type in other than
the last postion. If the function passes the screens then a funcNode for
the function is added to the Container's slice of nodes.

The implementaton does not yet support the edges on the graph that are
implicit in the parameters and non-error return types of the function
added with AddFunc.