Implemented From for Container with AddFunc called.
A Container now maintains two maps from types to a slice of nodes. There
is one map from a types.Type to the nodes that provide that type
(providedBy) and another map from a types.Type to the nodes that require
that type (requiredBy).

AddFunc() adds the new funcNode to the requiredBy map for the types
required by the node and adds the node to the providedBy map for the
types provided by the node. From() uses the two maps to find nodes
whose required types are provided by the node given to From().