Refactored rootNode into a normal node in Container.
The singleton nature of rootNode is not longer built into the rootNode
type itself. It is now enfored by the SetRoot() method on Container. A
rootNode is added to the 'nodes' collection in a Container and its
provided and required types get entries added to 'requiredBy' and
'providedBy' (though rootNode has an empty provides()).

This change in the treatment of rootNode allows much type speceific
special cases for rootNode to be removed. Code for rootNode and funcNode
can now be collapsed into a common case for the graph.Directed methods
(though this has only been done for Has(), Nodes(), and From() so
far).

The renaming of the 'node' interface to 'commonNode' is to prevent name
clashes with a common parameter name for the graph.Directed methods
(i.e. From(node graph.Node) ).