Refactored missingNode to be a normal node.
The singleton nature of the missingNode is enforced by the Container,
and not by having special type handling for missingNode. This simplifes
Container.From() to a single type of logic.

The missingNode now provides() the types that are required by some other
node but are not provided by any other node. This is computed
dynamically each time it is needed.