Added root auto-detection to AddFunc.
Container.AddFunc() will now check the provided types for the funcNode
it adds to see if any of them are root types and will set the
Container's root to that type if one of them is. It is an error to set a
second root type either on a single call to AddFunc() or on a subsequent
call.

A root type for the purposes of auto-detection is a type whose method
set includes a nullary method named "Run". It may be useful in the
future to add different tests for auto-detecting a root type.